Opt frame listen

+++ build_module/service.sh
@@ -7,5 +7,10 @@ until [ -d "/sdcard/Android" ]; do
 	sleep 1
 done
 
+if lsmod | grep -q perfmgr_mtk; then
+	touch $MODDIR/disable
+	exit
+fi
+
 killall fas-rs
 nohup $MODDIR/fas-rs >/dev/null 2>&1 &
